fix: have lock issues workflow be generated by Projen
This required refactoring the project setup a bit to change it to
TypeScript and allow it to import the already-existing workflow
being used by the individual prebuilt provider repos. There might
possibly be some issues since Projen does not make it easy to
transition from JS to TS once the project already exists.