chore: implement Ansgar's solution for licenses through Projen
This brings the license generation into the Projen template to
ensure that it does not get overwritten in the future, copying
Ansgar's implementation from the AWS Adapter. This solution is not
entirely complete in that it doesn't address how/where to get the
creation year for the individual prebuilt provider repos; I wasn't
sure whether we should build that functionality in the repo-manager
or handle it here, or whether to hard-code it into e.g. a JSON map
or whether to make calls to the GitHub API to look up the repo's
creation date.