chore: randomize the time when lock workflow runs
I've been getting a lot of notifications about the lock workflow hitting rate limits which I think is happening because every prebuilt provider is trying to run this workflow at the same time. This randomizes the time, which I assume should lead to the different repos running it at different times.