chore: upgrade dependencies for providers weekly instead of nightly (#328)
This can help cut down on the GitHub notification spam as well as the
excessive Mergify usage. Most of the time, Projen is the only thing that
gets updated anyway, and we really don't need those updates nightly.

The downside is that this also updates CDKTF patch versions, but if we
ever critically need to update all providers with a patch, I suppose we
could just trigger it manually from cdktf-repository-manager.