chore: automate upgrading Node.js on cdktf-provider-project (#337)
This uses the same script that's already been tested & proven to work
well on hashicorp/cdktf-aws-cdk. Note that this only upgrades Node on
the provider-project itself because the Node version used by the
individual provider repos is managed by cdktf-repository-manager
instead.

Also adds the automerge and auto-approve workflows we've been using
elsewhere. We don't _have_ to start using them, but this way these
automations are there if/when we need them.

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>