chore: use "fix" in changelogs to make important updates more visible (#343)
It's been bothering me for some time that some of the most critical
updates in our changelogs are super buried. Here's a particularly
egregious example:

![Screen Shot 2023-10-04 at 14 39
06](https://github.com/cdktf/cdktf-provider-project/assets/855115/776f1152-24e2-4d11-92db-7385eb6f1195)

The really important part here is the provider version change, but it's
buried so deep at the end of a long list of routine Projen updates that
you might miss it.

IIRC `standard-version` orders changelogs feat > fix > chore, so a
simple solution is to use the "fix" type here to put the provider
version at the start of the changelog instead of at the end. It also
feels like more than a chore since upgrading a provider could come with
new or improved functionality.

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>