chore(ci): address two minor workflow issues (#438)
Quick fixes for two small issues:

- the `cancel-in-progress: true` on the automerge and auto-approve
workflows was sometimes (not consistently) causing _all_ runs of a
workflow to be canceled, instead of forcing only the most recent one to
run. This feels like a bug on GitHub's side but whatever, we should be
able to fix it by removing that setting.
- for whatever reason the alert-open-prs workflow has been failing a lot
this week with 503s/504s, which shows up as red on our dashboard. If
this workflow happens to fail once, it's not that big of a deal, so I
applied the same fix that we used for the lock workflow --
`continue-on-error: true` to have the workflow not show up as failed.