fix(ci): remove invalid properties from force-release workflow (#442)
Unfortunately, the rollout of #437 failed because it turns out there was
some invalid syntax in the `force-release` workflow, which the
`action-validator` package was failing the build on. This highlighted
one flaw of this setup: because the `force-release` workflow doesn't
exist in this project, and only in the prebuilt providers / implementers
of this project, it wasn't being run through the validation at
build/test time here in the source, and so it's possible to write a bad
workflow here, publish a release, and break all implementing projects as
a result.

In order to be a good citizen, I not only fixed the issue with
`force-release` itself but also figured out how to add tests for all of
the workflows, again using `action-validator`. I manually verified that
with the bad syntax still included in `force-release.ts`, the tests were
failing (as they should). When I removed the bad syntax, all the tests
pass. Win!